requirements.txt
opencv-python
flask-opencv-streamer
ultralytics
deepface
bbox-visualizer
prettytable
flask
click>=8.0.0
PyYAML>=6.0